+++ CleanArchitechture-master/CleanArchitechture-master/CleanArchitecture.Application/Interfaces/IReviewServices.cs
@@ -10,7 +10,7 @@ namespace CleanArchitecture.Application.Interfaces
     {
         Task<List<object>> GetList_Reviews(int skip, int take, string str, long userID, CancellationToken cancellationToken = default);
         Task<ReviewsDto> Review_Create(Review review, CancellationToken cancellationToken = default);
-        Task<int> Review_Del(int reviewID, long userID, CancellationToken cancellationToken = default);
+        Task<long> Review_Del(int reviewID, long userID, CancellationToken cancellationToken = default);
         Task<List<Review>> GetList_Reviews_ByOwner(int reviewID, long userID, CancellationToken cancellationToken = default);
     }
 }
